Refactored monkey patching to server plugin api methods, moved some logic to separate files
