+++ plugin.rb
@@ -62,57 +62,10 @@ after_initialize do
     "events_required"
   ].each do |key|
     Site.preloaded_category_custom_fields << key if Site.respond_to? :preloaded_category_custom_fields
+    add_to_class(:category, key.to_sym) do
+      self.custom_fields[key] || SiteSetting.respond_to?(key) ? SiteSetting.send(key) : false
+    end
     add_to_serializer(:basic_category, key.to_sym) { object.send(key) }
-  end
-
-  class ::Category
-    def events_min_trust_to_create
-      if self.custom_fields['events_min_trust_to_create'].present?
-        self.custom_fields['events_min_trust_to_create'].to_i
-      else
-        SiteSetting.events_min_trust_to_create
-      end
-    end
-
-    def events_enabled
-      if self.custom_fields['events_enabled'] != nil
-        self.custom_fields['events_enabled']
-      else
-        SiteSetting.events_enabled
-      end
-    end
-
-    def events_agenda_enabled
-      if self.custom_fields['events_agenda_enabled'] != nil
-        self.custom_fields['events_agenda_enabled']
-      else
-        SiteSetting.events_agenda_enabled
-      end
-    end
-
-    def events_calendar_enabled
-      if self.custom_fields['events_calendar_enabled'] != nil
-        self.custom_fields['events_calendar_enabled']
-      else
-        SiteSetting.events_calendar_enabled
-      end
-    end
-
-    def events_required
-      if self.custom_fields['events_required'] != nil
-        self.custom_fields['events_required']
-      else
-        false
-      end
-    end
-
-    def events_event_label_no_text
-      if self.custom_fields['events_event_label_no_text'] != nil
-        self.custom_fields['events_event_label_no_text']
-      else
-        false
-      end
-    end
   end
 
   module EventsSiteSettingExtension
@@ -125,8 +78,7 @@ after_initialize do
     end
   end
 
-  require_dependency 'site_settings/type_supervisor'
-  class SiteSettings::TypeSupervisor
+  class ::SiteSettings::TypeSupervisor
     prepend EventsSiteSettingExtension
   end
 
@@ -154,106 +106,79 @@ after_initialize do
   load File.expand_path('../controllers/api_keys.rb', __FILE__)
 
   # a combined hash with iso8601 dates is easier to work with
-  require_dependency 'topic'
-  class ::Topic
-    def has_event?
-      self.custom_fields['event_start'].present? &&
-      self.custom_fields['event_start'].is_a?(Numeric) &&
-      self.custom_fields['event_start'] != 0
+  add_to_class(:topic, :has_event?) do
+    self.custom_fields['event_start'].present? &&
+    self.custom_fields['event_start'].is_a?(Numeric) &&
+    self.custom_fields['event_start'] != 0
+  end
+
+  add_to_class(:topic, :event) do
+    return nil unless has_event?
+    event = { start: Time.at(custom_fields['event_start']).iso8601 }
+
+    if custom_fields['event_end']&.nonzero?
+      event[:end] = Time.at(custom_fields['event_end']).iso8601
     end
 
-    def event
-      return nil unless has_event?
-      event = { start: Time.at(custom_fields['event_start']).iso8601 }
-
-      if custom_fields['event_end']&.nonzero?
-        event[:end] = Time.at(custom_fields['event_end']).iso8601
-      end
-
-      if custom_fields['event_timezone'].present?
-        event[:timezone] = custom_fields['event_timezone']
-      end
-
-      if custom_fields['event_all_day'].present?
-        event[:all_day] = custom_fields['event_all_day']
-      end
-
-      if custom_fields['event_version'].present?
-        event[:version] = custom_fields['event_version']
-      end
-
-      if event_rsvp
-        event[:rsvp] = event_rsvp
-
-        if event_going_max
-          event[:going_max] = event_going_max
-        end
-
-        if event_going
-          event[:going] = User.find(event_going).pluck(:username)
-        end
-      end
-
-      event
+    if custom_fields['event_timezone'].present?
+      event[:timezone] = custom_fields['event_timezone']
     end
 
-    def event_going
-      if self.custom_fields['event_going']
-        self.custom_fields['event_going']
-      else
-        false
+    if custom_fields['event_all_day'].present?
+      event[:all_day] = custom_fields['event_all_day']
+    end
+
+    if custom_fields['event_version'].present?
+      event[:version] = custom_fields['event_version']
+    end
+
+    if event_rsvp
+      event[:rsvp] = event_rsvp
+
+      if event_going_max
+        event[:going_max] = event_going_max
+      end
+
+      if event_going
+        event[:going] = User.find(event_going).pluck(:username)
       end
     end
 
-    def event_rsvp
-      if self.custom_fields['event_rsvp'] != nil
-        self.custom_fields['event_rsvp']
-      else
-        false
-      end
-    end
+    event
+  end
 
-    def event_going_max
-      if self.custom_fields['event_going_max'].to_i > 0
-        self.custom_fields['event_going_max'].to_i
-      else
-        nil
-      end
+  [
+    "event_going",
+    "event_rsvp",
+    "event_going_max",
+  ].each do |key|
+    add_to_class(:topic, key.to_sym) do
+      self.custom_fields[key] || false
     end
   end
 
-  require_dependency 'topic_view_serializer'
-  class ::TopicViewSerializer
-    attributes :event
-
-    def event
-      object.topic.event
-    end
-
-    def include_event?
-      object.topic.has_event?
-    end
+  add_to_serializer(:topic_view, :event, false) do
+    object.topic.event
   end
 
-  require_dependency 'topic_list_item_serializer'
-  class ::TopicListItemSerializer
-    attributes :event, :event_going_total
+  add_to_serializer(:topic_view, :include_event?, false) do
+    object.topic.has_event?
+  end
 
-    def event
-      object.event
-    end
+  add_to_serializer(:topic_list_item, :event, false) do
+    object.event
+  end
 
-    def include_event?
-      object.has_event?
-    end
+  add_to_serializer(:topic_list_item, :include_event?, false) do
+    object.has_event?
+  end
 
-    def event_going_total
-      object.event_going ? object.event_going.length : 0
-    end
+  add_to_serializer(:topic_list_item, :event_going_total) do
+    object.event_going ? object.event_going.length : 0
+  end
 
-    def include_event_going_total?
-      include_event?
-    end
+  add_to_serializer(:topic_list_item, :include_event_going_total?) do
+    include_event?
   end
 
   User.register_custom_field_type('calendar_first_day_week', :integer)
@@ -310,80 +235,24 @@ after_initialize do
       end
     end
 
-  DiscourseEvent.on(:post_created) do |post, opts, user|
-    is_wizard_event = !!(opts[:topic_opts] && opts[:topic_opts][:custom_fields] && opts[:topic_opts][:custom_fields]['event'])
+  load File.expand_path('../lib/event_creator.rb', __FILE__)
 
-    if post.is_first_post? && (opts[:event] || is_wizard_event)
-      topic = Topic.find(post.topic_id)
-      event_params = is_wizard_event ? opts[:topic_opts][:custom_fields]['event'] : opts[:event]
-      guardian = Guardian.new(user)
-      guardian.ensure_can_create_event!(topic.category)
-      event = event_params.is_a?(String) ? ::JSON.parse(event_params) : event_params
-      event_start = event['start']
-      event_end = event['end']
-      event_all_day = event['all_day']
-      timezone = event['timezone']
-      rsvp = event['rsvp']
-      going_max = event['going_max']
-      going = event['going']
-      event_version = 1
-
-      topic.custom_fields['event_start'] = event_start.to_datetime.to_i if event_start
-      topic.custom_fields['event_end'] = event_end.to_datetime.to_i if event_end
-      topic.custom_fields['event_all_day'] = event_all_day === 'true' if event_all_day
-      topic.custom_fields['event_timezone'] = timezone if timezone
-      topic.custom_fields['event_rsvp'] = rsvp if rsvp
-      topic.custom_fields['event_going_max'] = going_max if going_max
-      topic.custom_fields['event_going'] = User.where(username: going).pluck(:id) if going
-      topic.custom_fields['event_version'] = event_version if event_version
-
-      topic.save_custom_fields(true)
-    end
+  on(:post_created) do |post, opts, user|
+    event_creator = ::EventCreator.new(post, opts, user)
+    event_creator.create
   end
 
-  TopicQuery.add_custom_filter(:start) do |topics, query|
-    if query.options[:start]
-      topics.where("topics.id in (
-        SELECT topic_id FROM topic_custom_fields
-        WHERE (name = 'event_start' OR name = 'event_end')
-        AND value >= '#{query.options[:start].to_datetime.beginning_of_day.to_i}'
-      )")
-    else
-      topics
-    end
+  add_to_class(:guardian, :can_create_event?) do |category|
+    category.events_enabled &&
+    can_create_topic_on_category?(category) &&
+    (is_staff? ||
+    (user && user.trust_level >= category.events_min_trust_to_create))
   end
 
-  TopicQuery.add_custom_filter(:end) do |topics, query|
-    if query.options[:end]
-      topics.where("topics.id in (
-        SELECT topic_id FROM topic_custom_fields
-        WHERE (name = 'event_start' OR name = 'event_end')
-        AND value <= '#{query.options[:end].to_datetime.end_of_day.to_i}'
-      )")
-    else
-      topics
-    end
+  add_to_class(:guardian, :can_edit_event?) do |category|
+    can_create_event?(category)
   end
 
-  module EventsGuardian
-    def can_create_event?(category)
-      category.events_enabled &&
-      can_create_topic_on_category?(category) &&
-      (is_staff? ||
-      (user && user.trust_level >= category.events_min_trust_to_create))
-    end
-
-    def can_edit_event?(category)
-      can_create_event?(category)
-    end
-  end
-
-  require_dependency 'guardian'
-  class ::Guardian
-    include EventsGuardian
-  end
-
-  require_dependency 'topic'
   class ::Topic
     attr_accessor :include_excerpt
   end
@@ -394,225 +263,12 @@ after_initialize do
     end
   end
 
-  require_dependency 'listable_topic_serializer'
   class ::ListableTopicSerializer
     prepend ListableTopicSerializerExtension
   end
 
-  require_dependency 'topic_query'
-  class ::TopicQuery
-    def list_agenda
-      @options[:unordered] = true
-      @options[:list] = 'agenda'
-
-      opts = {
-        remove_past: SiteSetting.events_remove_past_from_agenda
-      }
-
-      opts[:status] = 'open' if SiteSetting.events_agenda_filter_closed
-
-      create_list(:agenda, {}, event_results(opts))
-    end
-
-    def list_calendar
-      @options[:unordered] = true
-      @options[:list] = 'calendar'
-
-      opts = {
-        limit: false,
-        include_excerpt: true,
-        remove_past: SiteSetting.events_remove_past_from_calendar
-      }
-
-      opts[:status] = 'open' if SiteSetting.events_calendar_filter_closed
-
-      create_list(:calendar, {}, event_results(opts))
-    end
-
-    def event_results(options = {})
-      topics = default_results(options.reverse_merge(ascending: 'true'))
-        .joins("INNER JOIN topic_custom_fields
-                ON topic_custom_fields.topic_id = topics.id
-                AND topic_custom_fields.name = 'event_start'
-                AND topic_custom_fields.value <> ''")
-
-      CalendarEvents::List.sorted_filters.each do |filter|
-        topics = filter[:block].call(topics, @options)
-      end
-
-      if options[:remove_past]
-        topics = topics.where("topics.id in (
-          SELECT topic_id FROM topic_custom_fields
-          WHERE (name = 'event_start' OR name ='event_end')
-          AND value > '#{Time.now.to_i}'
-        )")
-      end
-
-      topics = topics.order("(
-          SELECT CASE
-          WHEN EXISTS (
-            SELECT true FROM topic_custom_fields tcf
-            WHERE tcf.topic_id::integer = topics.id::integer
-            AND tcf.name = 'event_start' LIMIT 1
-          )
-          THEN (
-            SELECT value::integer FROM topic_custom_fields tcf
-            WHERE tcf.topic_id::integer = topics.id::integer
-            AND tcf.name = 'event_start' LIMIT 1
-          )
-          ELSE 0 END
-        ) ASC")
-
-      if options[:include_excerpt]
-        topics.each { |t| t.include_excerpt = true }
-      end
-
-      topics
-    end
-  end
-  
-  module ListControllerEventsExtension
-    USER_API_KEY ||= "user_api_key"
-    USER_API_CLIENT_ID ||= "user_api_client_id"
-    # Logging in with user API keys normally only works by passing certain headers.
-    # As we cannot force third-party software to send those headers, we need to fake
-    # them using request parameters.
-    def current_user
-      if params.key?(USER_API_KEY)
-        request.env[Auth::DefaultCurrentUserProvider::USER_API_KEY] = params[USER_API_KEY]
-        if params.key?(USER_API_CLIENT_ID)
-          request.env[Auth::DefaultCurrentUserProvider::USER_API_CLIENT_ID] = params[USER_API_CLIENT_ID]
-        end
-      end
-      super
-    end
-  end
-
-  require 'icalendar/tzinfo'
-  class ::ListController
-    skip_before_action :ensure_logged_in, only: [:calendar_ics, :calendar_feed]
-    skip_before_action :set_category, only: [
-      :agenda_feed,
-      :calendar_ics,
-      :calendar_feed,
-    ]
-    
-    def agenda_feed
-      self.send('event_ics', name: 'agenda')
-    end
-
-    def calendar_feed
-      self.send('event_feed', name: 'calendar')
-    end
-
-    def calendar_ics
-      self.send('event_ics', name: 'calendar')
-    end
-    
-    def agenda_feed_category
-      self.send('event_feed', name: 'agenda')
-    end
-    
-    def calendar_feed_category
-      self.send('event_feed', name: 'calendar')
-    end
-
-    def calendar_ics_category
-      self.send('event_ics', name: 'calendar')
-    end
-
-    def event_feed(opts = {})
-      discourse_expires_in 1.minute
-
-      guardian.ensure_can_see!(@category) if @category
-
-      title_prefix = @category ? "#{SiteSetting.title} - #{@category.name}" : SiteSetting.title
-      base_url = @category ? @category.url : Discourse.base_url
-      list_opts = {}
-      list_opts[:category] = @category.id if @category
-
-      @title = "#{title_prefix} #{I18n.t("rss_description.events")}"
-      @link = "#{base_url}/#{opts[:name]}"
-      @atom_link = "#{base_url}/#{opts[:name]}.rss"
-      @description = I18n.t("rss_description.events")
-      @topic_list = TopicQuery.new(nil, list_opts).list_agenda
-
-      render 'list', formats: [:rss]
-    end
-
-    def event_ics(opts = {})      
-      guardian.ensure_can_see!(@category) if @category
-      
-      name_prefix = @category ? "#{SiteSetting.title} - #{@category.name}" : SiteSetting.title
-      base_url = @category ? @category.url : Discourse.base_url
-
-      calendar_name = "#{name_prefix} #{I18n.t("webcal_description.events")}"
-      calendar_url = "#{base_url}/calendar"
-      list_opts = {}
-      list_opts[:category] = @category.id if @category
-      list_opts[:tags] = params[:tags] if params[:tags]
-
-      if current_user &&
-         SiteSetting.respond_to?(:assign_enabled) &&
-         SiteSetting.assign_enabled
-        list_opts[:assigned] = current_user.username if params[:assigned]
-      end
-
-      tzid = params[:time_zone] || ( SiteSetting.respond_to?(:events_timezone_default) && SiteSetting.events_timezone_default.present? && SiteSetting.events_timezone_default ) || "Etc/UTC"
-      tz = TZInfo::Timezone.get tzid
-
-      cal = Icalendar::Calendar.new
-      cal.x_wr_calname = calendar_name
-      cal.x_wr_timezone = tzid
-      # add timezone once per calendar
-      event_now = DateTime.now
-      timezone = tz.ical_timezone event_now
-      cal.add_timezone timezone
-
-      @topic_list = TopicQuery.new(current_user, list_opts).list_calendar
-
-      @topic_list.topics.each do |t|
-        if t.event && t.event[:start]
-          event = CalendarEvents::Helper.localize_event(t.event, tzid)
-          timezone = tz.ical_timezone event[:start]
-
-          ## to do: check if working later
-          if event[:format] == :date_only
-            event[:start] = event[:start].to_date.strftime "%Y%m%d"
-            event[:end] = (event[:end].to_date+1).strftime "%Y%m%d" if event[:end]
-          end
-
-          if event[:going].present?
-            going_emails = User.where(username: event[:going]).map(&:email)
-          end
-
-          cal.event do |e|
-            e.dtstart = Icalendar::Values::DateOrDateTime.new(event[:start], 'tzid' => tzid).call
-            if event[:end]
-              e.dtend = Icalendar::Values::DateOrDateTime.new(event[:end], 'tzid' => tzid).call
-            end
-            e.summary = t.title
-            e.description = t.url << "\n\n" << t.excerpt #add url to event body
-            e.url = t.url #most calendar clients don't display this field
-            e.uid = t.id.to_s + "@" + Discourse.base_url.sub(/^https?\:\/\/(www.)?/,'')
-            e.sequence = event[:version]
-            
-            if going_emails
-              going_emails.each do |email|
-                e.append_attendee "mailto:#{email}"
-              end
-            end
-          end
-        end
-      end
-      
-      cal.publish
-
-      render body: cal.to_ical, formats: [:ics], content_type: Mime::Type.lookup("text/calendar") unless performed?
-    end
-    
-    prepend ListControllerEventsExtension
-  end
+  load File.expand_path('../lib/topic_query_edits.rb', __FILE__)
+  load File.expand_path('../controllers/list_controller_edits.rb', __FILE__)
 
   on(:approved_post) do |reviewable, post|
     event = reviewable.payload['event']
